CRUD operations added to api endpoints

+++ app/Http/Controllers/RegistrationController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use App\Meeting;
+use App\User;
 
 class RegistrationController extends Controller
 {
@@ -23,28 +25,34 @@ class RegistrationController extends Controller
         $meeting_id = $request->input('meeting_id');
         $user_id = $request->input('user_id');
 
-        $meeting = [
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
+        $meeting = Meeting::findOrFail((int)$meeting_id);
+        $user = User::findOrFail((int)$user_id);
+
+        $message = [
+            'message' => 'User is already registered for the meeting',
+            'user' => $user,
+            'meeting' => $meeting,
+            'unregister' => [
+                'href' => 'api/v1/meeting/registration/' . $meeting->id,
+                'method' => 'DELETE'
             ]
         ];
 
-        $user = [
-            'name' => 'Name'
-        ];
+        // If the user already exists for this meeting
+        if ($meeting->users()->where('users.id', $user->id)->first()) {
+            return response()->json($message, 404);
+        }
+
+        $user->meetings()->attach($meeting);
 
         $response = [
             'message' => 'User successfully registered for meeting',
             'meeting' => $meeting,
             'user' => $user,
             'unregister' => [
-                'href' => 'api/v1/meeting/registration/1',
+                'href' => 'api/v1/meeting/registration/' . $meeting->id,
                 'method' => 'DELETE'
-            ];
+            ]
         ];
 
         return response()->json($response, 201);
@@ -58,31 +66,21 @@ class RegistrationController extends Controller
      */
     public function destroy($id)
     {
-        $meeting = [
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
-        ];
-
-        $user = [
-            'name' => 'Name'
-        ];
+        
+        $meeting = Meeting::findOrFail((int)$id);
+        $meeting->users()->detach();
 
         $response = [
             'message' => 'User successfully unregistered for meeting',
             'meeting' => $meeting,
-            'user' => $user,
+            'user' => 'tbd',
             'register' => [
                 'href' => 'api/v1/meeting/registration',
                 'method' => 'POST',
                 'params' => 'user_id, meeting_id'
-            ];
+            ]
         ];
 
-        return response()->json($response, 200)
+        return response()->json($response, 200);
     }
 }
